+++ stream-chat-android-compose-sample/src/main/java/io/getstream/chat/android/compose/sample/data/PredefinedUserCredentials.kt
@@ -187,6 +187,16 @@ object PredefinedUserCredentials {
         UserCredentials(
             apiKey = API_KEY,
             user = User(
+                id = "pvelikov",
+                name = "Petar Velikov",
+                image = "https://ca.slack-edge.com/T02RM6X6B-U07LDJZRUTG-a4129fed05b6-512",
+            ),
+            token = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoicHZlbGlrb3YifQ." +
+                "d5eenuTIZD5gZh7rHiv3lYbE8uOqUiHfwULtUr8a-l0",
+        ),
+        UserCredentials(
+            apiKey = API_KEY,
+            user = User(
                 id = "qatest1",
                 name = "QA Test 1",
                 image = "https://getstream.imgix.net/images/random_svg/QT.png",