+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/messages/MessagesScreen.kt
@@ -699,6 +699,7 @@ public fun BoxScope.AttachmentsPickerMenu(
                             } else {
                                 1
                             },
+                            enforceUniqueVote = action.switches.none { it.key == "maxVotesAllowed" && it.enabled },
                         ),
                     )
                 }