+++ stream-chat-android-ui-components-sample/src/demo/kotlin/io/getstream/chat/ui/sample/application/AppConfig.kt
@@ -155,6 +155,14 @@ object AppConfig {
         ),
         SampleUser(
             apiKey = apiKey,
+            id = "pvelikov",
+            name = "Petar Velikov",
+            image = "https://ca.slack-edge.com/T02RM6X6B-U07LDJZRUTG-a4129fed05b6-512",
+            token = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoicHZlbGlrb3YifQ.d5eenuTIZD5gZh7rHiv3lYbE8uOqUiHfwULtUr8a-l0",
+            language = "en",
+        ),
+        SampleUser(
+            apiKey = apiKey,
             id = "qatest1",
             name = "QA Test 1",
             image = "https://getstream.imgix.net/images/random_svg/QT.png",